Use react-loadable to deferring loading LineGraph.js
